--- test/cli/ambiguous-this-module-include/ambiguous_this_module_include.rb
@@ -1,17 +1,0 @@
-# typed: true
-
-module Parent
-end
-
-module Child
-  include Parent
-end
-
-module A
-  extend T::Sig
-  sig {params(parent: Parent).void}
-  def example(parent)
-    if parent.is_a?(Child)
-    end
-  end
-end
