--- test/cli/ambiguous-this-module-include/test.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-set -e
-
-main/sorbet --silence-dev-message test/cli/ambiguous-this-module-include/ 2>&1
-
