+++ test/cli/suggest-kernel/suggest-kernel.rb
@@ -13,3 +13,19 @@ sig { params(x: Foo).void }
 def example(x)
   x.is_a?(Foo)
 end
+
+module Parent
+end
+
+module Child
+  include Parent
+end
+
+module A
+  extend T::Sig
+  sig {params(parent: Parent).void}
+  def example(parent)
+    if parent.is_a?(Child)
+    end
+  end
+end
